+++ source/programs/instruments/instrumentResample.cpp
@@ -2,7 +2,7 @@
 /**
 * @file instrumentResample.cpp
 *
-* @brief Resample data to given time series using polynomial prediction or least squares polynomial fit.
+* @brief Resample data to given time series.
 *
 * @author Andreas Kvas
 * @author Sebastian Strasser
@@ -15,27 +15,8 @@
 #define DOCSTRING docstring
 static const char *docstring = R"(
 This program resamples \file{instrument data}{instrument} to a given
-\configClass{timeSeries}{timeSeriesType} using a resampling \config{method} out of the following:
-
-\begin{itemize}
-\item \config{polynomial}: Polynomial prediction using a moving polynomial of \config{polynomialDegree}.
-  The optimal polynomial is chosen based on the centricity of the data points around the resampling
-  point and the distance to all polynomial data points. All polynomial data points must be within
-  \config{maxDataPointRange}. Resampling points within \config{maxExtrapolationDistance} of the
-  polynomial will be extrapolated.
-
-  \fig{!hb}{0.5}{instrumentResample_polynomial}{fig:instrumentResample_polynomial}{Example of polynomial prediction when resampling from 5 to 1 minute sampling}
-
-\item \config{leastSquaresPolynomialFit}. A polynomial of \config{polynomialDegree} is estimated using
-  all data points within \config{maxDataPointDistance} of the resampling point. This polynomial is then
-  used to predict the resampling point. A resampling point will be extrapolated if there are only data
-  points before/after as long as the closest one is within \config{maxExtrapolationDistance}.
-
-  \fig{!hb}{0.5}{instrumentResample_leastSquares}{fig:instrumentResample_leastSquares}{Example of least squares polynomial fit when resampling from 5 to 1 minute sampling}
-\end{itemize}
-
-The elements \config{maxDataPointRange}, \config{maxDataPointDistance}, and \config{maxExtrapolationDistance}
-are given in the unit of seconds. If negative values are used, the unit is relative to the median input sampling.
+\configClass{timeSeries}{timeSeriesType} using a resampling
+\configClass{method}{interpolatorTimeSeriesType}.
 
 This program can also be used to reduce the sampling of an instrument file,
 but a better way to reduce the sampling of noisy data with regular sampling
@@ -46,13 +27,13 @@ out the data with \program{InstrumentReduceSampling}.
 /***********************************************/
 
 #include "programs/program.h"
-#include "base/polynomial.h"
 #include "files/fileInstrument.h"
+#include "classes/interpolatorTimeSeries/interpolatorTimeSeries.h"
 #include "classes/timeSeries/timeSeries.h"
 
 /***** CLASS ***********************************/
 
-/** @brief Resample data to given time series using polynomial prediction or least squares polynomial fit.
+/** @brief Resample data to given time series.
 * @ingroup programsGroup */
 class InstrumentResample
 {
@@ -60,7 +41,7 @@ public:
   void run(Config &config, Parallel::CommunicatorPtr comm);
 };
 
-GROOPS_REGISTER_PROGRAM(InstrumentResample, SINGLEPROCESS, "Resample data to given time series using polynomial prediction or least squares polynomial fit.", Instrument)
+GROOPS_REGISTER_PROGRAM(InstrumentResample, SINGLEPROCESS, "Resample data to given time series", Instrument)
 
 /***********************************************/
 
@@ -68,34 +49,14 @@ void InstrumentResample::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
 {
   try
   {
-    FileName      outName, inName;
-    UInt          polynomialDegree;
-    Double        maxDataPointRange, maxExtrapolationDistance;
-    TimeSeriesPtr timeSeries;
-    Bool          isLeastSquares = FALSE;
-    std::string   choice;
+    FileName                  outName, inName;
+    InterpolatorTimeSeriesPtr interpolator;
+    TimeSeriesPtr             timeSeries;
 
-    readConfig(config, "outputfileInstrument", outName, Config::MUSTSET,  "", "");
-    readConfig(config, "inputfileInstrument",  inName,  Config::MUSTSET,  "", "");
-    if(readConfigChoice(config, "method", choice, Config::MUSTSET,  "polynomial", "resampling method"))
-    {
-      if(readConfigChoiceElement(config, "polynomial", choice, "polynomial prediction"))
-      {
-        isLeastSquares = FALSE;
-        readConfig(config, "polynomialDegree",         polynomialDegree,         Config::MUSTSET, "3",        "degree of the moving polynomial");
-        readConfig(config, "maxDataPointRange",        maxDataPointRange,        Config::MUSTSET, "-(3+1.1)", "[seconds] all degree+1 data points must be within this range for a valid polynomial");
-        readConfig(config, "maxExtrapolationDistance", maxExtrapolationDistance, Config::DEFAULT, "-1.1",     "[seconds] resampling points within this range of the polynomial will be extrapolated");
-      }
-      if(readConfigChoiceElement(config, "leastSquaresPolynomialFit", choice, "least squares polynomial fit"))
-      {
-        isLeastSquares = TRUE;
-        readConfig(config, "polynomialDegree",         polynomialDegree,         Config::MUSTSET, "",  "degree of the estimated polynomial");
-        readConfig(config, "maxDataPointDistance",     maxDataPointRange,        Config::MUSTSET, "",  "[seconds] all data points within this distance around the resampling point will be used");
-        readConfig(config, "maxExtrapolationDistance", maxExtrapolationDistance, Config::DEFAULT, "0", "[seconds] resampling points within this range of the polynomial will be extrapolated");
-      }
-      endChoice(config);
-    }
-    readConfig(config, "timeSeries", timeSeries, Config::MUSTSET, "", "resampled points in time");
+    readConfig(config, "outputfileInstrument", outName,      Config::MUSTSET, "", "");
+    readConfig(config, "inputfileInstrument",  inName,       Config::MUSTSET, "", "");
+    readConfig(config, "method",               interpolator, Config::MUSTSET, "", "resampling method");
+    readConfig(config, "timeSeries",           timeSeries,   Config::MUSTSET, "", "resampled points in time");
     if(isCreateSchema(config)) return;
 
     logStatus<<"read instrument data <"<<inName<<">"<<Log::endl;
@@ -103,9 +64,8 @@ void InstrumentResample::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
     const std::vector<Time> timesNew = timeSeries->times();
 
     logStatus<<"resample data"<<Log::endl;
-    Polynomial polynomial;
-    polynomial.init(arc.times(), polynomialDegree, FALSE/*throwException*/, isLeastSquares, maxDataPointRange, maxExtrapolationDistance);
-    const Matrix A = polynomial.interpolate(timesNew, arc.matrix().column(1, Epoch::dataCount(arc.getType(), TRUE)));
+    interpolator->init(arc.times(), FALSE/*throwException*/);
+    const Matrix A = interpolator->interpolate(timesNew, arc.matrix().column(1, Epoch::dataCount(arc.getType(), TRUE)));
 
     Arc arcNew;
     Epoch *epoch = Epoch::create(arc.getType());